Align Orders screen ownership and naming
